+++ RELEASES.md
@@ -11,6 +11,7 @@ Version 1.2.2 (unreleased)
   * The missing function
     <code>[gmp][gmp-1-2]::[MPZ\_ROINIT\_N][gmp-mrn-1-2]</code> was
     added.
+  * The missing macro [`MPFR_DECL_INIT`][mpfr-di-1-2] was added.
 
 Version 1.2.1 (2020-03-13)
 ==========================
@@ -35,6 +36,7 @@ Version 1.2.0 (2020-01-18)
 [gmp-1-2]: https://docs.rs/gmp-mpfr-sys/~1.2/gmp_mpfr_sys/gmp/index.html
 [gmp-mrn-1-2]: https://docs.rs/gmp-mpfr-sys/~1.2/gmp_mpfr_sys/gmp/fn.MPZ_ROINIT_N.html
 [gmp-rs-1-2]: https://docs.rs/gmp-mpfr-sys/~1.2/gmp_mpfr_sys/gmp/struct.randstate_t.html
+[mpfr-di-1-2]: https://docs.rs/gmp-mpfr-sys/~1.2/gmp_mpfr_sys/macro.MPFR_DECL_INIT.html
 
 Version 1.1.14 (2019-07-04)
 ===========================